pyproject.toml
[tool.poetry]
name = "markbench-tests"
version = "1.0.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.10"
pywinauto = "0.6.8"
pyyaml = "6.0"
python-dotenv = "0.20.0"
pyfiglet = "0.8.post1"
PyAutoGUI = "^0.9.53"
PyDirectInput = "^1.0.4"
opencv-python = "^4.5.5"
Pillow = "^9.1.1"
psutil = "^5.9.1"
pandas = "^1.4.2"
imutils = "^0.5.4"
matplotlib = "^3.5.2"
protobuf = "^4.21.2"
WMI = "^1.5.1"
mss = "^7.0.1"
requests = "^2.28.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
coverage = "^6.3.3"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
black = "^23.9.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# https://docs.pytest.org/en/6.2.x/reference.html#ini-options-ref
[tool.pytest.ini_options]
testpaths = ["./tests"]

[tool.isort]
profile = "black"
line_length = 88

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'